Add files via upload
Cattroller is a new idea and proof of concepts for creating an IoT based device, which can be integrated with block-chain based video games. in this approach it'll attract video games fans who like breeding and adopting digital games but afraid to invest in block-chains implementations by giving them the eager to try this remarkable experience, and at the same time which can also stay in their pocket.